Make the controller aware of mirroring configuration
Motivation
Disconnected environments have mirroring or mapping set using ICSP and
that works as expected for the container runtime on the hosts. This
doesn't work when manually trying to extract the release image from the
controller.

Modification
Use the github.com/containers/image and related libraris. They support
registries.conf mirroring configuration, and also very easy to work
with.

Signed-off-by: Roy Golan <rgolan@redhat.com>