Add label to oauth route
In the managed scenario (Service Delivery), a separate router shard
(IngressController) can be created to make control plane routes public,
making it possible to keep most routes in the cluster private.
In a PublicAndPrivate use case, the only route in a HyperShift control
plane that needs to be public is the OAuth route.

This commit adds a label to oauth routes so that if you have a dedicated
router shard for them, you are able to use a selector in your router
shard to only select oauth routes.