E2E: Use zaptest logger for cluster creation
This change makes us use the zaptest logger for cluster creation in E2E
tests. The zaptest logger scopes its output to the current testcase,
which should help a lot with legibility of E2E output. In order for this
change to be useful, all providers need to use the passed logger for
their cluster creation log, this change only does that for AWS and
Azure.