TestAutoscaling: Fix to work on Azure
Azure uses a different naming scheme for the AZs label value
($region-$zone), so instead of using the input value we try to find a
node for the nodepool we care about and use its zone value.

Also adds some basic validation to the test inputs to avoid running it
in configurations that won't allow it to succeed.