Add configmap containing supported versions to hypershift namespace
Introduces a controller that maintains a configmap in the `hypershift`
namespace that contains a list of supported minor versions by the
current controller.

This is done by a controller and not laid down by the installer because
the operator itself is the source of truth for what it supports, and not
the CLI.

The ConfigMap will be used by the MCE UI to filter the versions that can
be used when creating a HostedCluster.