Socks 5 proxy: Only use mgmt cluster for cloud apis for ingress operator
The socks5 proxy has code to avoid sending traffic destined for cloud
provider apis through konnektivity in order to not require nodes to be
up for that communication to work.

When running a management cluster with a proxy setup and not setting
proxy env vars in the socks5 proxy, that code effectively means that it
is not possible to communicate with those apis, as the sidecar can not
route it as it has no direct Internet connectivity.

The above currently breaks imagestreamimports if the import is from a docker
registry hosted in AWS S3, as the Openshift APIServer is supposed to
route all its imports through Konnektivity, which is why its proxy
sidecar does not get the mgmt cluster proxy env vars.

The only reason for special casing cloud provider apis here is the
ingress operator, as it should be able to setup DNS before nodes exist
in order to speed up cluster creation.

As this special case is surprising and an easy footgun, hide it behind a
flag and only activate this for the Ingress Operator.

This change should stabilize the proxy conformance periodic, which has
currently a high failure rate which ultimately always boilds down to
being unable to import the `tools` image in the `openshift` namespace,
which is used by many tests.