ensure optional mount always added to ibmcloud kms pod for appropriate container build
There is a problem in the unmanaged auth path where an optional mount that is necessary for the build out of the container is not getting mounted. It is only getting mounted in the managed auth path. This ensures the volume is added in both paths so unamanged authentication (customer provided authentication) works