Move high-level cert secret reconciliation funcs into support/certs
Currently, there is the support/certs package that provides relatively
low-level functions for generating certificates. On top of that we have
higher level functions in
control-plane-operator/controllers/hostedcontrolplane/pki that provide
cert and ca secret reconciliation logic. This logic is not specific to
the CPO and needed in other places.

This change moves the CA and signed cert secret reconciliation logic
into support/certs, which leaves only the code for specific cert
configuration and tests in the pki package.

This change is a pure refactor that should not be changing anything and
precondition to start using the same certs/pki code everywhere.