Correctly manage infra status condition when there is no LB controller
We currently only update the infra status after we got an event. If we
reconcile a hostedcluster that uses an LB for the apiserver in a
management cluster that does not support those, we will never update the
status because we never get an event.

Recheck after a minute when the infrastructure is not ready to make sure
the condition gets updated regularly.