E2E: Fix flakes due to inability of finding matching nodepools
The E2E tests currently try to match nodepools to a hostedcluster by
using an ownerReference on it. This ownerreference is added by the
nodepoool controller and doesn't exist initially, which makes this
approach racy and thus causes flakes.

Use the same matching logic as the nodepool controller
(`nodepool.spec.clusterName`) to avoid this.