Fix clusters that use LB expose strategy with a custom port
In https://github.com/openshift/hypershift/pull/1614 the KAS was changed
to always listen on port 6443 in oder to support privileged external
ports by doing the mapping external port -> internal port in the service
object.

This resulted in an invalid KAS networkpolicy if the external port is
not 6443, as it would allowlist the external port rather than 6443 where
the KAS is actually listening.