Add a drain controller for inplace upgrades
This attempts to add a inplace drainer controller to the HCCO which
mirrors the functionality in self-driving OCP's MCO to allow
nodes to be drained before performing inplace upgrades.

This drainer acts separately from the inplace upgrader, only watching
node objects in the hosted cluster. When it detects a node annotation
requesting drain (from the MCD running on the node), the controller
cordons and drains the node. The node is uncordoned post upgrade via a
similar annotation.

This is not an optimal solution. Longer term we should probably consider
having a central draining/reboot coordination mechanism.

Some TODOs:
1. properly handle nodepool drain timeouts
2. properly handle error messages and status