Add node controller to HCCO Manager
What:
This introduces a feature to reconcile Nodes with an opinionated Hypershift label pointing to the owning NodePool.

Why:
This is useful to list and filter Nodes by NodePool. This is useful for all operators to have a single and consistent bidirectional way for NodePool<->Node without having to reimplement the same logic, vendoring clients and consuming compute resources. E.g NTO tuneD.

How:
Change the NodePool controller to propagate the NodePool annotation down to Machines.
Introduce a Node controller within the HCCO manager:
- Watches Nodes.
- Find the Machine by using the appropriate CAPI annotations.
- Find the NodePool annotation.
- Apply the annotation as a Label into the Node.

In future we might extend this to support label propagation into Nodes via NodePool API.