Oauth: Fix hostname when using DNS indirection and no public router LB
When DNS indirection is used but the cluster doesn't have a public LB
for its private router, the logic incorrectly set the default ingress
domain on the oauth route. This results in clients using the service DNS
domain name -> hitting the router -> getting a PKI failure because the
router doesn't know anything about this hostname and falls through to
its 404 page, which uses a different cert.