Use kubeconfig port for worker haproxy
We are currently only using a default port for the Kube APIServer in the
public case (unless it's overriden). This doesn't take into
consideration a nodeport in the case that the APIServer is exposed via
NodePort. A safer configuration route is to take the port of the
external kubeconfig which already has the calculated external endpoint
of the Kube APIServer.

When using private link, we can keep using what we had before because
the port is fixed.