Ensure generated junit.xml has a defined suite name
gotestsum's generated junit.xml doesn't have a suite name, and the test
names themselves are really generic sounding, e.g. 'TestAutoscaling'.
Adding a suite name will help us identify them in Sippy as belonging to
hypershift.

I couldn't find a way with gotestsum to get it to put something useful
in the suite name, so this just hacks it in with sed.  Open to better
solutions.