Make zone spread only apply within a given revision
We currently apply zone spread for all revisions for a given workload.
This means that a new revision can only be rolled out after a replica of
the old revision was removed.

This PR fixes that by:
* Moving the calculation of the zone spread affinity into the ApplyTo
  funcs
* Calculating a hash of the podTemplate there
* Applying a label to the pod with that hash
* Using all pod labels including the one with the hash in the
  AntiAffinity rule

As a sideeffect, this removes the requirement to know the pods labels by
the time DeploymentConfig.SetDefaults() is called. In many cases, they
weren't known by that time so it was called with a nil labelmap, which
resulted in the Zone spread code being short-circuited. With this
change, everything that calls SetDefaults and has more than one replica
will get the zone spread, as it is makes sense for all components.

I manually verified that with this change and a
sufficiently-sized management cluster, a HA controlplane upgrade results
in zero failed scheduling events.