bug: Fixes HOSTEDCP-528, OIDCConfigurationInvalidReason should be always present
A side-scenario bug could happen when someone or some piece deletes the OIDC configuration while it's in use.
The race condition ends on the AWS Instance deleted and the platform provider cannot continue with the
AWS components deletion.
For that we've set the default status as false the ValidOIDCConfiguration status condition,
then always will come from false state and rechecked in a posteriory step of the reconcile loop.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>