bug: HOSTEDCP-569 HC controller blocked reconcile verifying releaseImage
This bug happens when you try to provision a HC and you have not created the PullSecret. The reconcile process will try to pull the ReleaseImage and verify it, but it cannot because cannot recover the Pull Secret.
With the current verification process the controller returns 'nil' and after 7 tries, the reconcile process gets blocked. The fix validates this situation of the missing pull secret before this second check, but in this case the controller returns an error, which made possible the controller to wait indefinitelly until the PullSecret gets created by the user.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>