NodePool API: rename spec.tunedConfig to spec.tuningConfig
This is needed to keep the name general because we will likely
use this field to reference PerformanceProfiles later on when
we enable the PAO / PerformanceProfile functionality (~4.13 timeframe).

As discussed in the HyperShift project sync meeting, we would like to
make the change while the NodePool API is still in alpha.