[Test Only] [HyperShift/KubeVirt] Reduce flakiness for NoticePreemptionOrFailedScheduling
HyperShift/KubeVirt's TestCreateCluster suite is running just after successfully deploying CNV and installing HyperShift.
At this point, virt-handler might not be fully operational and not register itself yet, which might lead to a transient failure in which
the cdi-importer pods could not be scheduled.
This PR adds an exception for the FailedScheduling or Preempted events originated from cdi-importer pods, as this is transient.

Note: We rely solely on the pod name, that it is prefixed with importer-, because at the time the events on the hostedcluster namespace are
being checked, the importer pod is gone, and we can't verify by pod labels at this point.

Signed-off-by: Oren Cohen <ocohen@redhat.com>