Pin PreviousReleaseImage
When the CPO is at N and the NodePool.spec.release at N-1 we fail to render ignition payload because openshift/machine-config-operator#3286 broke backward compatibility. This PRs lets it fall back to MCO previously supported flags.
This commit ping the image used in the upgrade test to a no broken version.