Honor MaxUnvailable for inplace upgrades (#1656)
* Honor MaxUnvailable for inplace upgrades

* Propagate inplace upgrade nodepool maxUnavailable

Add an annotation to pass maxUnavailable between nodepool controller to
HCCO via machineset, like we do for other nodepool configs.

The nodepool controller will take an int or percent value and convert it
to a number of nodes maxUnavailable to pass via the annotation. We
should definitely convert this to a better interface in the future, but
for now, this is the easiest way to do so.

Add a unit test for handling the new maxUnavailable for InPlace. We should
eventually also beef up our e2e for actual upgrade testing.

Co-authored-by: Yu Qi Zhang <jerzhang@redhat.com>