inplaceupgrader: handle multiple upgrade versions
Compared to self-driving OCP, there is an additional scenario to
consider: since the ConfigMap contents are synced separately, those
will change on the fly in the pod, meaning that we could have a
scenario where there are multiple queue'ed updates, in which case the
MCD will just jump straight to the latest version.

This will cause the MCD to softlock, so let's make sure for those
unavailable nodes, we are also update their desired configuration.