Added capability to WaitFor(Un)ReadyNodes to filter by NodePool
This capability give some powerfull function to check the nodes which belongs to a NodePool without using the NodePool object itself.
we just need to add as an optional parameter an slice of strings as las argument in the function call.
Warning: For now we only supports the first element of the slice as NodePool name, but in the future we will include also the rest of them.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>