[HyperShift/KubeVirt] Test on CNV 4.12 from nightly channel
In order to test hypershift with kubevirt provider on OCP 4.12, we'll need to use CNV 4.12 as well since it is compatible with Pod Security Admission (PSA) enforcement.
CNV 4.12 is available to be consumed from the nightly channel at quay.io/openshift-cnv, and it holds the most latest build that passed QE verification.

Signed-off-by: Oren Cohen <ocohen@redhat.com>