Add annotation to allow resource requests overrides
Introduces an annotation that if set on the HostedCluster resource, can
override the default resource requests on control plane deployments.
This is needed in CI where we know beforehand that we need a particular
amount of memory and CPU for the Kube APIServer of management clusters.
Setting this will enable the root cluster to scale compute
appropriately.