AUTH-323: separate the cluster-distributed client-ca from the rootCA (#1837)
* cpo/c/hcp/*: Separate Aggregate Client trust chain

Signed-off-by: Krzysztof Ostrowski <kostrows@redhat.com>

* add helper to pki manifests not to copy-paste secret meta all the time

* fix aggregator cert EKU

* Add the kube-controlplane signers

Also fixes the aggregator-related function names from
"aggregate*" to "aggregator*" to avoid confusion

* add KAS->kubelet signer

* add system-admin signer and fix admin client cert EKUs

* add the CSR signer and the total client CA bundle

* remove "combined-ca" and fix bootstrapping

This CA combined CAs that are supposed to be used for both user
and server authentication. That is wrong, you want these two be
separate so that if you are an application that wants to choose
to trust server certificates, you don't automatically enable
client cert authentication from these signers, too (and vice versa).

* cosmetic: kas_pki_setup: remove the need to add the constant ownerref

* fix the client CA trust for all the components using it

* fix SA kubeconfig generation

The kubeconfig client certs are now being signed by the CSR signer.

Previously, the SAs were missing the namespace-specific group. This is
now fixed by using the upstream functions to get SA name/groups.

Also removes the invalid ServerAuth EKU from the client certs.

* go mod tidy && go mod vendor

* sign the metrics client cert with the CSR signer

* fix MCS server cert EKU

* allow KCM create CMs in kube-system for leader-election

* fix router-controller-manager with the shared client ca

* fix user logins - sign the openshift authenticator by the CSR signer

This also adds the rootCA to the kubeconfig used by the KAS to
authenticate itself to the oauth-apiserver in order for it to be
able to verify the oauth-apiserver serving cert.

* fixup: return the CSR signer to original location for clean upgrade tests

* temporarily add the rootCA to the aggregated and client CA bundles

This is to get the upgrade tests green. This commit should be
reverted once the control-plane version we're upgrading from
in the test has all the client certificates signed by the
correct signers.

Signed-off-by: Krzysztof Ostrowski <kostrows@redhat.com>
Co-authored-by: Krzysztof Ostrowski <kostrows@redhat.com>