Optionally use rhobs.monitoring group for monitoring resources
This change enables the CLI, hypershift operator, and control plane
operator to optionally use the rhobs.monitoring/v1 group version when
creating and reconciling monitoring resources. This is a global setting
that must be set on the HyperShift  operator and gets propagated to all
hosted control planes managed by the operator.

This will enable the HyperShift operator to create resources that can be
understood by the RHOBS monitoring stack.