WIP: HOSTEDCP-582: Add CEL validations to hostedcluster v1beta1
This uses CEL validations available in kubernetes 1.25 and above,
requiring OpenShift 4.12 or greater.  I have tested with OpenShift 4.10
and while the validations do not work, they are also ignored so it should
be safe to have them in place for all versions.

Note that I am not setting maxLength here for the strings. I looked
through the source code and it seems for simple string comparisons we
should be OK here. I'd really rather not have to set this if we don't have
to. However if we run into issues we'll have to think about setting it.

Still a WIP as I need to do more testing.