Add ValidIdentityProvider condition
There's scenarios where the identity provider might have been conflicted, e.g. delete by accident.
In that case the HC is not able to operate successfully e.g. aws endpoints, ec2 machines... and there's no clear signal in the HC.
This is also true if the HC is in deleting state, which will hang forever with no clear signal.

This PR tries to allivites those scenarios by capturing them in a condition.