Prevent cluster deletion hangs by using heartbeat condition
If the hcco is not available, the cpo will wait indefinitely for it to
update hcp conditions to indicate that guest resources have been
deleted. To solve that, the hcco will periodically update a condition to
let the CPO that it's working on deletion. If the condition is not
updated after a certain amount of time, the CPO gives up waiting and
proceeds with the rest of the deletion.