Enable creating infra for AWS Outpost through the cli
The PR adds support for creating the nodes on AWS Outpost.
The motivation for this change is to enables users who wish to run the nodes and their workloads
on their premises, either for privacy or to enhance the AWS Outpost.
Creating the private subnet in AWS Outpost will lead to creating the instances on it as well.

The option to specify the AWS Outpost ARN is added to infra create command and to cluster create
command by flag `--outpost-arn $OUTPOST_ARN`

AWS Outpost has several limitations:
* AWS Outpost supports only gp2 for root volume type. The unsupported value will be overridden by gp2.
  * More info - https://docs.aws.amazon.com/whitepapers/latest/aws-outposts-high-availability-design/storage.html
* AWS Outpost is homed to a specific availability zone. Multi AZ will not be supported.
  The zone should match the AWS Outpost AZ. The Zone is determined according to AWS Outpost AZ if not provided.

Signed-off-by: Moti Asayag <masayag@redhat.com>