Set missing internal route label for ovnkube-sbdb
Right now in a private setup the ovnkube-sbdb route is never addmited as it's missing the label that let the ingress controller reconcile it.
This was most likely missed here https://github.com/openshift/hypershift/pull/1861/files.
The e2e included in this PR should prevent this from happening again.