Expose proxy, identityProviders and limited support metrics
This extend existing metrics exposed by the HO to include Expose proxy, identityProviders and limited support metrics. These metrics are useful to infer symptons and point to causes when observing a fleet of clusters. E.g. Lets say we notice a subset of customers having IDP issues but no reported outages on github. We would be able to analyze the metrics and indeed say that github was having issues if we had outages fleetwide across github.
These metrics are used in ROSA today. In future if they grow in a very opinionated way we might want to decouple them from the HO.