Enable creating infra for AWS with Local Gateway
In order to allow the user to associate a VPC with AWS Outpost local
gateway, few steps needs to be added to the infrastructure creation:
1. Associate a VPC with the Local Gateway (it is expected to be only one
   per outpost).
2. Add a route from the private subnet to the local gateway.

In addition, when destroying an infrastructure, there is a need to
diassociate the VPC from the local gateway route table. This step was
added to the infra destory command.

Signed-off-by: Moti Asayag <masayag@redhat.com>