Add CSR signer to Kubelet client CA
The CA bundle that the kubelet uses to verify mTLS certificates needs to
include the CA used to sign the KAS client certs to allow communication
from the KAS as well as the CSR signer CA to allow service accounts that
have the right RBAC to access metrics.

This commit adds a new CA bundle (kubelet-client-ca) that includes both
the KASToKubelet signer as well as the CSR signer. This allows the
prometheus-k8s service account to scrape metrics from the kubelet.