Fix infrastructure.status.apiServerInternalURI for PublicAndPrivate
Currently in a PublicAndPrivate setup behind a proxy, nodes remain unready because ovn-node pods fail to talk to the control plane osdb and api endpoints.
This is because the connection to .hypershift.local tries to go through the proxy.
This PR set's the right value in the infra status that will result in hypershift.local being added to the NO_PROXY list of the ovn-pods via
https://github.com/openshift/cluster-network-operator/pull/1692/files#diff-689ddfc925eea62478b3853bc2468faffd84032b887d3798d5e99bd9a54d0b55R77
https://github.com/openshift/cluster-network-operator/pull/1692/files#diff-689ddfc925eea62478b3853bc2468faffd84032b887d3798d5e99bd9a54d0b55R77

Aditionally we'll also let the CNO set .hypershift.local in NO_PROXY unconditionally as this is a design invariant https://github.com/openshift/cluster-network-operator/pull/1692