Create default security group for AWS clusters
Enables the control plane operator to create a default security group
for AWS HostedClusters. If a NodePool does not specify a security group,
the default security group is used for its AWSMachineTemplate.

NodePool reconciliation will wait for the security group to be present
before proceeding with template creation. For HostedClusters with a
control plane operator that does not implement the default security
group creation, NodePool reconciliation will proceed as before.

Security group creation status is reported as a condition on the
HostedCluster. The security group ID is reported in a platform-specific
status of the HostedCluster.