Update the pull secret source for ignition payload
Update the pull secret source for ignition payload. The NodePool controller will save a hash of the pull secret in every newly created token secret. During creation of an ignition payload, a comparison between the token secret's pull secret hash and a hash of the pull secret will be made to ensure they match. The pull secret will be written to the bootstrap config directory, overwriting the one from the MCS ConfigMap.

Signed-off-by: Bryan Cox <brcox@redhat.com>