OCPBUGS-8231: Fix cleanup of volumes on cluster deletion
The current code that updates the CloudResourcesDestroyed condition on
the hostedcontrolplane resource is not updating the LastTransitionTime.
The hostedcontrolplane controller understands this to mean that the hcco
is no longer active and proceeds to remove the finalizer from the
hostedcontrolplane on the configured timeout. This fix ensures that the
LastTransitionTime gets updated on the condition so it can behave as a
heartbeat for the hostedcontrolplane controller.