OCPBUGS-8381: Use appropriate serving certificate for OAuth
We currently only copy the self-signed cert to openshift-config-managed
of the hosted cluster. However, if a certificate has been specified for
the oauth endpoint via spec.config.apiServer, the bundle that we copy
over should contain the additional certificate.

Adds CA bundle to the oauth server configuration that contains certs
specified through apiserver namedcerts and the self-signed cert that we
normally generate.