HOSTEDCP-801: Expose external DNS for private cluster endpoints
For clusters with an endpoint access of type Private, that use route as
a publishing strategy and specify a hostname, the hostname is exposed
externally via external-dns using a service of type ExternalName.

Clusters can be switched back and forth between Private and
PublicAndPrivate and the external-dns records are adjusted as needed.