ensure well known public domains do not get proxied on image imports
In client environments that run with a HTTP_PROXY: that proxy information is lost when the image import is done on the openshift-apiserver side and sent through the konnectivity-sidecar. There are a set of well known domains that are necessary for the MVP set of samples that get deployed in an openshift cluster. These well known public domains should not have to be proxied into the client environment and instead can go direct from the management cluster. Any other defined domains do need to proxy into the cluster network for the ability to access on prem registries.