OCPBUGS-10492: Abandon HCP cleanup after a timeout
Sets up a timeout for hostedcontrolplane deletion after which any
pending cleanup is abandoned and the finalizer on the hostedcontrolplane
is removed.

This will prevent a hostedcluster delete from hanging forever if there is
something preventing the deletion of resources it needs to cleanup.
Currently the timeout is set at 3 hrs.