Preserve false status of ValidAWSIdentityProvider condition
When the hostedcontrolplane has been removed, we may still depend on the
ValidAWSIdentityProvider condition to determine whether to abandon the
AWSEndpointService. Current behavior is that the condition status
reverts to Unknown. This prevents the AWSEndpointService controller from
abandoning the cleanup of the VPC endpoint.

This change preserves the value of the condition if it was false and can
no longer be determined via the hostedcontrolplane resource.