W/A for KubeVirt provider with RHCOS 9.2
Due to changes in RHCOS 9.2, the ignition config is not able to be fetched with the openstack provider.
However, it works with kubevirt ignition provider.
Adding a workaround that uses a modified RHCOS containerdisk that sets the kernel command line parameter of 'ignition.platform.id=kubevirt' instead of 'ignition.platform.id=openstack'

Signed-off-by: Oren Cohen <ocohen@redhat.com>