properly reconcile with user specified changes for in proxy configuration
Users are able to and directed to configure the cluster-wide proxy directly in the guest cluster when needing to setup local proxying in the guest cluster network: https://docs.openshift.com/container-platform/4.10/networking/enable-cluster-wide-proxy.html. Currently: even when the proxy configuration in the hosted cluster is nil: we are squashing all these changes and revoking what they are doing locally. This makes it impossible for them to configure the proxy within their network (they should not need to have to proxy these configurations to the hostedcluster resource each time and get it to propogate as that can be controlled by a separate entity like a managed cloud provider). This allows the user to create those changes and have them persist