Move cluster deletion duration metric into controller
Atm we have high frequency reconciliation loop where we constantly review the over all state of the world by looping over all clusters.
It's worth noting that With this particular change we'll be capturing something different i.e. "when we remove all finalisers we own", which doesn't necessarily is the same than "when the HC is removed from etcd".