Configurable SRE MetricsSet
Introduces the use of a ConfigMap to configure the metrics relabelings
to use in control plane ServiceMonitors. The ConfigMap should live in
the hypershift operator namespace, be named 'sre-metric-set' and have a
single key named 'config'. The contents of the config should be a YAML
file with keys for each of the control plane components that produces a
service monitor. An example is included in support/metrics/testdata.

Changes to the configuration ConfigMap result in the ConfigMap getting
copied to all hosted control planes and the Control Plane Operator pod
in those namespaces to be restarted. For the CPO pod, the configuration
is only read once during startup and used by the service monitor
reconciliation functions.