Health check load balancers only on public clusters
The health check on internal load balancers will only work if the load
balancer includes the private subnets. However, this is not guaranteed
unless the subnets of the management cluster are labeled with public or
private labels. For now, we'll only check the public load balancers in a
non-private cluster and just the KAS service in a private cluster.