HOSTEDCP-947: Maxed the QUOTA_BACKEND_BYTES to the recommended limit
As the ETCD metioned in the documentation, the max size for this quota is 8G, above that they will raise a performance warning. This limit will only prevent the ETCD to write over 8G in the PV, but the PV default size is not modified, because the PVC size could be raised directly.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>