Refactored the metrics recollection in HO to be funciontally testable
This refactor involves the package creation of metrics which is now a
child package from the hostedcluster controller one. Now metrics are
accesible from other packages.

The metrics recollection now are summoned using their own function name, some
new functions was created for this purpose and also their own tests.

Also the TestMetrics function was moved to a multiple ones, covering all
the new content included in the PR.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>