KubeVirt: Handle deletion of the cache DV on an edge case
The credetionals for the external infra clusters are stored in the
controller memory. If for some reason, the pod is been restarted while
deleting a nodepool, there is a chnace tat we'll loose this information,
and we won't be able to delete the cache DataVolume.

This PR fixes it by copying the missing information from the
HostedCluster CR, to the NodePool's status, so it will be available even
after restart.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>