Use different ports for MCS in the ignition provider
There is a restriction on querying ports 22623 and 22624 on the
cluster's supported ip families everywhere in the cluster. This lead to
every request to MCS resulting in `connection refused` in a dual-stack
cluster.

An IPv4-only cluster got around this issue because the temp MCS server
bound to the ports on both IPv4 and IPv6 and the request resolved
`localhost` to `::1`. This allowed the request to get past the blocks on
the ports on the IPv4 familiy.

On a dual-stack cluster the ports are blocked for both IPv4 and IPv6 so
the request could never succeed.

Resolves https://issues.redhat.com/browse/OCPBUGS-12972