NTO: Enable PerformanceProfile reconcile
Modify the code for NTO Tuned manifests reconciliation so it could also
reconcile when there is a PerformanceProfile in the ConfigMap instead of
a Tuned.

Read all the configmaps referenced in NodePool.spec.TunedConfig field
and propagate any PerformanceProfile manifest into ConfigMaps in the
hosted control plane namespace where NTO, and so PerformanceProfile
controller, is running and can read them.