use combined ca for controller manager arg to appropriately set ca in service accounts pod use
There was a change on 4.12 and above that removed the combined-ca and instead moved to using the root-ca for the ca bundle of the cloud controller manager arg. This means the ca pods use in their service accounts does not contain the cluster signer ca and therefore has no local trust bundle to be able to validate tls connections if making calls to the kubelet server apis. This fix restores that behavior that existed in 4.11 and below by using the associated ca bundle that has that info in the later releases

Reference to 4.11 branch: https://github.com/openshift/hypershift/blob/release-4.11/control-plane-operator/controllers/hostedcontrolplane/kcm/deployment.go#L33