Enable CVO to evaluate conditional updates on self-managed HyperShift deployed on OpenShift
This commit will enable the Cluster Version Operator (CVO) to evaluate
conditional updates on self-managed HyperShift deployed on an OpenShift
management cluster.

For the CVO to evaluate conditional updates, it accesses a PromQL query
service holding the appropriate information. In the standalone
OpenShift, it is the thanos-querier service in the openshift-monitoring
namespace.

This commit will point the hosted CVO at the mentioned service when
specified via a command line flag as well. To use the thanos-querier
service, commit utilizes its tenancy port 9092. This port allows the
users to query metrics from a specified namespace (in this case, the
hosted CVO can query metrics from its hosted control plane namespace).
For the authorization, the hosted CVO is given a new Role with the
minimal needed permissions to be able to query its HCP metrics.

For the Control Plane Operator to be able to give this permission to the
CVO, it needs this permission as well. The same case applies to the
HyperShift Operator (HSO).

When installing HyperShift, the
`--enable-cvo-management-cluster-metrics-access` flag can be now passed
to modify specific API objects accordingly to allow the hosted CVO to
access the HCP metrics. The flag is not supported when
`--rhobs-monitoring` is set, as the hosted CVO is unable to access
metrics scraped by RHOBS objects at the moment. In the case, that the
flag `--enable-cvo-management-cluster-metrics-access` is set on a
non-OpenShift management cluster, the CVO will continue to report
`Recommended = Unknown` for conditional updates.

A new network policy is created to only allow egress communication to
the PromQL query service pod's port from the hosted CVO pods.